release images when a tag is pushed to the git repo
Pushing a git tag will trigger a PipelineRun which will collect all the
images created by builds associated with the commit the tag was applied
to, and create a "override" Snapshot. This Snapshot will be tested using
the "bats-integration" test, and on successful completion the images in
the Snapshot will be pushed to their respective repos in the
quay.io/ramalama org. Each image will have two tags applied, one matching
the commit sha the image was built from, and the other matching the git
tag that was pushed.